refactor(loginToHome): Remove login route and move that logic to home page, replacing what was previously there. Remove Firebase analytics. Add explanation comments to useStaffHover.
